Backend/backend_project/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
axios==0.4.0
bcrypt==4.1.2
bitarray==2.9.2
black==24.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.0
cytoolz==0.12.3
Django==5.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
ecdsa==0.18.0
eth-account==0.11.0
eth-hash==0.6.0
eth-keyfile==0.8.0
eth-keys==0.5.0
eth-rlp==1.0.1
eth-typing==4.0.0
eth-utils==4.0.0
eth_abi==5.0.0
frozenlist==1.4.1
hexbytes==0.3.1
idna==3.6
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lxml==5.1.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
packaging==23.2
parsimonious==0.9.0
pathspec==0.12.1
platformdirs==4.2.0
protobuf==4.25.3
py-solc-x==2.0.2
pycparser==2.21
pycryptodome==3.20.0
Pygments==2.17.2
PyJWT==2.8.0
pyotp==2.9.0
pytz==2023.3.post1
pyunormalize==15.1.0
pywin32==306
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rlp==4.0.0
rpds-py==0.18.0
six==1.16.0
sqlparse==0.4.4
toolz==0.12.1
typing_extensions==4.10.0
tzdata==2023.4
urllib3==2.1.0
web3==6.15.1
websockets==12.0
yarl==1.9.4
lru-dict==1.2.0